pyproject.toml
[project]
name = "regeneracao"
version = "0.1.0"
description = ""
authors = [
    {name = "Pedro Henrique Windisch",email = "pedro@windisch.com.br"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "earthengine-api (>=1.7.16,<2.0.0)",
    "geopandas (>=1.1.2,<2.0.0)",
    "pandas (>=3.0.1,<4.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
